requirements.txt
selenium==4.15.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-telegram-bot==20.7
schedule==1.2.0
python-dotenv==1.0.0
webdriver-manager==4.0.1
aiohttp==3.9.1
asyncio==3.4.3
pandas==2.1.4
sqlalchemy==2.0.23